SonaDogru

+++ Veresiye.UI/FrmCompanies.cs
@@ -14,15 +14,23 @@ namespace Veresiye.UI
     public partial class FrmCompanies : Form
     {
         private readonly ICompanyService companyService;
-        public FrmAdd frmAdd;
-        public FrmEdit frmEdit;
+        private readonly FrmAdd frmAdd;
+        private readonly FrmCompanyEdit frmCompanyEdit;
         //private readonly FrmRegister frmRegister;
-        public FrmCompanies(CompanyService companyService, FrmAdd frmAdd, FrmEdit frmEdit)
+        public FrmCompanies(ICompanyService companyService, FrmAdd frmAdd, FrmCompanyEdit frmCompanyEdit)
         {
-         
+            
             this.companyService = companyService;
+            this.frmAdd = frmAdd;
+           this.frmCompanyEdit = frmCompanyEdit;
          
             InitializeComponent();
+           
+            this.frmAdd.MasterForm = this;
+            this.frmAdd.MdiParent = this.MdiParent;
+
+            this.frmCompanyEdit.MasterForm = this;
+            this.frmCompanyEdit.MdiParent = this.MdiParent;
         }
 
         private void FrmCompanies_Load(object sender, EventArgs e)
@@ -37,21 +45,55 @@ namespace Veresiye.UI
         }
 
         private void BtnDelete_Click(object sender, EventArgs e)
-        { 
-            int id = int.Parse(dataGridView1.SelectedRows[0].Cells[0].Value.ToString());
-            companyService.Delete(id);
-            LoadGrid();
-        }
-
-
-        public void LoadGrid()
         {
-            dataGridView1.DataSource = companyService.GetAll();
+           
+
+             
+
+            if (this.dataGridView1.SelectedRows.Count>0)
+            { var cevap = MessageBox.Show("Silmek istediğinize emin misiniz?", "Dİkkat!", MessageBoxButtons.YesNo);
+                if (cevap == DialogResult.Yes)
+                {
+                int id = int.Parse(dataGridView1.SelectedRows[0].Cells[0].Value.ToString());
+                companyService.Delete(id);
+                LoadCompanies();
+                }
+                else
+                {
+                    MessageBox.Show("Silme işlemi gerçekleştirilmiyor");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Lütfen silmek istediğiniz Firmayı seçiniz!");
+            }
+           
         }
+
 
         private void BtnAdd_Click(object sender, EventArgs e)
         {
+            frmAdd.Show();
+            this.frmAdd.ClearForm();//form açılırken temizlenmesi için!
             
+        }
+
+        private void BtnEdit_Click(object sender, EventArgs e)
+        {
+            if (this.dataGridView1.SelectedRows.Count>0)
+            {
+               
+               int selectedId=int.Parse( dataGridView1.SelectedRows[0].Cells[0].Value.ToString());
+                frmCompanyEdit.Show();
+                frmCompanyEdit.LoadForm(selectedId);
+              
+                
+            }
+            else
+            {
+                MessageBox.Show("Lütfen düzenlemek istediğiniz firmayı seçiniz");
+            }
+           
         }
     }
 }
